@@ -30,64 +30,64 @@
     "bcrypt-nodejs": "^0.0.3",
     "bluebird": "^3.5.1",
     "body-parser": "^1.18.2",
-    "compression": "^1.7.1",
-    "connect-mongo": "^2.0.0",
-    "dotenv": "^4.0.0",
+    "compression": "^1.7.2",
+    "connect-mongo": "^2.0.1",
+    "dotenv": "^5.0.1",
     "errorhandler": "^1.5.0",
-    "express": "^4.16.2",
-    "express-flash": "0.0.2",
+    "express": "^4.16.3",
+    "express-flash": "^0.0.2",
     "express-session": "^1.15.6",
-    "express-validator": "^4.3.0",
-    "fbgraph": "^1.4.1",
-    "lodash": "^4.17.4",
+    "express-validator": "^5.1.0",
+    "fbgraph": "^1.4.4",
+    "lodash": "^4.17.5",
     "lusca": "^1.5.2",
-    "mongoose": "^4.13.11",
-    "nodemailer": "^4.4.1",
+    "mongoose": "^5.0.12",
+    "nodemailer": "^4.6.4",
     "passport": "^0.4.0",
     "passport-facebook": "^2.1.1",
     "passport-local": "^1.0.0",
-    "pug": "^2.0.0-rc.4",
-    "request": "^2.83.0",
+    "pug": "^2.0.3",
+    "request": "^2.85.0",
     "request-promise": "^4.2.2",
-    "winston": "^2.4.0"
+    "winston": "^2.4.1"
   },
   "devDependencies": {
-    "@types/async": "^2.0.45",
+    "@types/async": "^2.0.48",
     "@types/bcrypt-nodejs": "^0.0.30",
     "@types/bluebird": "^3.5.20",
     "@types/body-parser": "^1.16.8",
-    "@types/compression": "^0.0.35",
+    "@types/compression": "^0.0.36",
     "@types/connect-mongo": "^0.0.35",
     "@types/dotenv": "^4.0.2",
     "@types/errorhandler": "^0.0.32",
     "@types/express": "^4.11.1",
     "@types/express-session": "^1.15.8",
-    "@types/jest": "^22.1.3",
-    "@types/jquery": "^3.2.17",
-    "@types/lodash": "^4.14.91",
+    "@types/jest": "^22.2.2",
+    "@types/jquery": "^3.3.1",
+    "@types/lodash": "^4.14.106",
     "@types/lusca": "^1.5.0",
-    "@types/mongodb": "^3.0.5",
-    "@types/mongoose": "^4.7.34",
+    "@types/mongodb": "^3.0.9",
+    "@types/mongoose": "^5.0.7",
     "@types/morgan": "^1.7.35",
-    "@types/node": "^9.4.6",
-    "@types/nodemailer": "^4.3.4",
+    "@types/node": "^9.6.1",
+    "@types/nodemailer": "^4.6.0",
     "@types/passport": "^0.4.3",
     "@types/passport-facebook": "^2.1.7",
     "@types/passport-local": "^1.0.32",
     "@types/request": "^2.47.0",
     "@types/shelljs": "^0.7.8",
     "@types/supertest": "^2.0.4",
-    "@types/winston": "^2.3.7",
+    "@types/winston": "^2.3.8",
     "chai": "^4.1.2",
     "concurrently": "^3.5.1",
-    "jest": "^22.0.4",
-    "node-sass": "^4.7.2",
-    "nodemon": "^1.13.0",
+    "jest": "^22.4.3",
+    "node-sass": "^4.8.3",
+    "nodemon": "^1.17.3",
     "shelljs": "^0.8.1",
     "supertest": "^3.0.0",
-    "ts-jest": "^22.0.4",
-    "ts-node": "^5.0.0",
+    "ts-jest": "^22.4.2",
+    "ts-node": "^5.0.1",
     "tslint": "^5.9.1",
-    "typescript": "^2.7.2"
+    "typescript": "^2.8.1"
   }
 }
