@@ -1,12 +1,12 @@
-import async from "async";
-import crypto from "crypto";
-import nodemailer from "nodemailer";
-import passport from "passport";
-import { default as User, UserModel, AuthToken } from "../models/User";
-import { Request, Response, NextFunction } from "express";
-import { IVerifyOptions } from "passport-local";
-import { WriteError } from "mongodb";
-const request = require("express-validator");
+import async from 'async';
+import crypto from 'crypto';
+import nodemailer from 'nodemailer';
+import passport from 'passport';
+import { default as User, UserModel, AuthToken } from '../models/User';
+import { Request, Response, NextFunction } from 'express';
+import { IVerifyOptions } from 'passport-local';
+import { WriteError } from 'mongodb';
+const request = require('express-validator');
 
 
 /**
@@ -15,10 +15,10 @@ const request = require("express-validator");
  */
 export let getLogin = (req: Request, res: Response) => {
   if (req.user) {
-    return res.redirect("/");
+    return res.redirect('/');
   }
-  res.render("account/login", {
-    title: "Login"
+  res.render('account/login', {
+    title: 'Login'
   });
 };
 
@@ -27,27 +27,27 @@ export let getLogin = (req: Request, res: Response) => {
  * Sign in using email and password.
  */
 export let postLogin = (req: Request, res: Response, next: NextFunction) => {
-  req.assert("email", "Email is not valid").isEmail();
-  req.assert("password", "Password cannot be blank").notEmpty();
-  req.sanitize("email").normalizeEmail({ gmail_remove_dots: false });
+  req.assert('email', 'Email is not valid').isEmail();
+  req.assert('password', 'Password cannot be blank').notEmpty();
+  req.sanitize('email').normalizeEmail({ gmail_remove_dots: false });
 
   const errors = req.validationErrors();
 
   if (errors) {
-    req.flash("errors", errors);
-    return res.redirect("/login");
+    req.flash('errors', errors);
+    return res.redirect('/login');
   }
 
-  passport.authenticate("local", (err: Error, user: UserModel, info: IVerifyOptions) => {
+  passport.authenticate('local', (err: Error, user: UserModel, info: IVerifyOptions) => {
     if (err) { return next(err); }
     if (!user) {
-      req.flash("errors", info.message);
-      return res.redirect("/login");
+      req.flash('errors', info.message);
+      return res.redirect('/login');
     }
     req.logIn(user, (err) => {
       if (err) { return next(err); }
-      req.flash("success", { msg: "Success! You are logged in." });
-      res.redirect(req.session.returnTo || "/");
+      req.flash('success', { msg: 'Success! You are logged in.' });
+      res.redirect(req.session.returnTo || '/');
     });
   })(req, res, next);
 };
@@ -58,7 +58,7 @@ export let postLogin = (req: Request, res: Response, next: NextFunction) => {
  */
 export let logout = (req: Request, res: Response) => {
   req.logout();
-  res.redirect("/");
+  res.redirect('/');
 };
 
 /**
@@ -67,10 +67,10 @@ export let logout = (req: Request, res: Response) => {
  */
 export let getSignup = (req: Request, res: Response) => {
   if (req.user) {
-    return res.redirect("/");
+    return res.redirect('/');
   }
-  res.render("account/signup", {
-    title: "Create Account"
+  res.render('account/signup', {
+    title: 'Create Account'
   });
 };
 
@@ -79,16 +79,16 @@ export let getSignup = (req: Request, res: Response) => {
  * Create a new local account.
  */
 export let postSignup = (req: Request, res: Response, next: NextFunction) => {
-  req.assert("email", "Email is not valid").isEmail();
-  req.assert("password", "Password must be at least 4 characters long").len({ min: 4 });
-  req.assert("confirmPassword", "Passwords do not match").equals(req.body.password);
-  req.sanitize("email").normalizeEmail({ gmail_remove_dots: false });
+  req.assert('email', 'Email is not valid').isEmail();
+  req.assert('password', 'Password must be at least 4 characters long').len({ min: 4 });
+  req.assert('confirmPassword', 'Passwords do not match').equals(req.body.password);
+  req.sanitize('email').normalizeEmail({ gmail_remove_dots: false });
 
   const errors = req.validationErrors();
 
   if (errors) {
-    req.flash("errors", errors);
-    return res.redirect("/signup");
+    req.flash('errors', errors);
+    return res.redirect('/signup');
   }
 
   const user = new User({
@@ -99,8 +99,8 @@ export let postSignup = (req: Request, res: Response, next: NextFunction) => {
   User.findOne({ email: req.body.email }, (err, existingUser) => {
     if (err) { return next(err); }
     if (existingUser) {
-      req.flash("errors", { msg: "Account with that email address already exists." });
-      return res.redirect("/signup");
+      req.flash('errors', { msg: 'Account with that email address already exists.' });
+      return res.redirect('/signup');
     }
     user.save((err) => {
       if (err) { return next(err); }
@@ -108,7 +108,7 @@ export let postSignup = (req: Request, res: Response, next: NextFunction) => {
         if (err) {
           return next(err);
         }
-        res.redirect("/");
+        res.redirect('/');
       });
     });
   });
@@ -119,8 +119,8 @@ export let postSignup = (req: Request, res: Response, next: NextFunction) => {
  * Profile page.
  */
 export let getAccount = (req: Request, res: Response) => {
-  res.render("account/profile", {
-    title: "Account Management"
+  res.render('account/profile', {
+    title: 'Account Management'
   });
 };
 
@@ -129,33 +129,33 @@ export let getAccount = (req: Request, res: Response) => {
  * Update profile information.
  */
 export let postUpdateProfile = (req: Request, res: Response, next: NextFunction) => {
-  req.assert("email", "Please enter a valid email address.").isEmail();
-  req.sanitize("email").normalizeEmail({ gmail_remove_dots: false });
+  req.assert('email', 'Please enter a valid email address.').isEmail();
+  req.sanitize('email').normalizeEmail({ gmail_remove_dots: false });
 
   const errors = req.validationErrors();
 
   if (errors) {
-    req.flash("errors", errors);
-    return res.redirect("/account");
+    req.flash('errors', errors);
+    return res.redirect('/account');
   }
 
   User.findById(req.user.id, (err, user: UserModel) => {
     if (err) { return next(err); }
-    user.email = req.body.email || "";
-    user.profile.name = req.body.name || "";
-    user.profile.gender = req.body.gender || "";
-    user.profile.location = req.body.location || "";
-    user.profile.website = req.body.website || "";
+    user.email = req.body.email || '';
+    user.profile.name = req.body.name || '';
+    user.profile.gender = req.body.gender || '';
+    user.profile.location = req.body.location || '';
+    user.profile.website = req.body.website || '';
     user.save((err: WriteError) => {
       if (err) {
         if (err.code === 11000) {
-          req.flash("errors", { msg: "The email address you have entered is already associated with an account." });
-          return res.redirect("/account");
+          req.flash('errors', { msg: 'The email address you have entered is already associated with an account.' });
+          return res.redirect('/account');
         }
         return next(err);
       }
-      req.flash("success", { msg: "Profile information has been updated." });
-      res.redirect("/account");
+      req.flash('success', { msg: 'Profile information has been updated.' });
+      res.redirect('/account');
     });
   });
 };
@@ -165,14 +165,14 @@ export let postUpdateProfile = (req: Request, res: Response, next: NextFunction)
  * Update current password.
  */
 export let postUpdatePassword = (req: Request, res: Response, next: NextFunction) => {
-  req.assert("password", "Password must be at least 4 characters long").len({ min: 4 });
-  req.assert("confirmPassword", "Passwords do not match").equals(req.body.password);
+  req.assert('password', 'Password must be at least 4 characters long').len({ min: 4 });
+  req.assert('confirmPassword', 'Passwords do not match').equals(req.body.password);
 
   const errors = req.validationErrors();
 
   if (errors) {
-    req.flash("errors", errors);
-    return res.redirect("/account");
+    req.flash('errors', errors);
+    return res.redirect('/account');
   }
 
   User.findById(req.user.id, (err, user: UserModel) => {
@@ -180,8 +180,8 @@ export let postUpdatePassword = (req: Request, res: Response, next: NextFunction
     user.password = req.body.password;
     user.save((err: WriteError) => {
       if (err) { return next(err); }
-      req.flash("success", { msg: "Password has been changed." });
-      res.redirect("/account");
+      req.flash('success', { msg: 'Password has been changed.' });
+      res.redirect('/account');
     });
   });
 };
@@ -194,8 +194,8 @@ export let postDeleteAccount = (req: Request, res: Response, next: NextFunction)
   User.remove({ _id: req.user.id }, (err) => {
     if (err) { return next(err); }
     req.logout();
-    req.flash("info", { msg: "Your account has been deleted." });
-    res.redirect("/");
+    req.flash('info', { msg: 'Your account has been deleted.' });
+    res.redirect('/');
   });
 };
 
@@ -211,8 +211,8 @@ export let getOauthUnlink = (req: Request, res: Response, next: NextFunction) =>
     user.tokens = user.tokens.filter((token: AuthToken) => token.kind !== provider);
     user.save((err: WriteError) => {
       if (err) { return next(err); }
-      req.flash("info", { msg: `${provider} account has been unlinked.` });
-      res.redirect("/account");
+      req.flash('info', { msg: `${provider} account has been unlinked.` });
+      res.redirect('/account');
     });
   });
 };
@@ -223,19 +223,19 @@ export let getOauthUnlink = (req: Request, res: Response, next: NextFunction) =>
  */
 export let getReset = (req: Request, res: Response, next: NextFunction) => {
   if (req.isAuthenticated()) {
-    return res.redirect("/");
+    return res.redirect('/');
   }
   User
     .findOne({ passwordResetToken: req.params.token })
-    .where("passwordResetExpires").gt(Date.now())
+    .where('passwordResetExpires').gt(Date.now())
     .exec((err, user) => {
       if (err) { return next(err); }
       if (!user) {
-        req.flash("errors", { msg: "Password reset token is invalid or has expired." });
-        return res.redirect("/forgot");
+        req.flash('errors', { msg: 'Password reset token is invalid or has expired.' });
+        return res.redirect('/forgot');
       }
-      res.render("account/reset", {
-        title: "Password Reset"
+      res.render('account/reset', {
+        title: 'Password Reset'
       });
     });
 };
@@ -245,26 +245,26 @@ export let getReset = (req: Request, res: Response, next: NextFunction) => {
  * Process the reset password request.
  */
 export let postReset = (req: Request, res: Response, next: NextFunction) => {
-  req.assert("password", "Password must be at least 4 characters long.").len({ min: 4 });
-  req.assert("confirm", "Passwords must match.").equals(req.body.password);
+  req.assert('password', 'Password must be at least 4 characters long.').len({ min: 4 });
+  req.assert('confirm', 'Passwords must match.').equals(req.body.password);
 
   const errors = req.validationErrors();
 
   if (errors) {
-    req.flash("errors", errors);
-    return res.redirect("back");
+    req.flash('errors', errors);
+    return res.redirect('back');
   }
 
   async.waterfall([
     function resetPassword(done: Function) {
       User
         .findOne({ passwordResetToken: req.params.token })
-        .where("passwordResetExpires").gt(Date.now())
+        .where('passwordResetExpires').gt(Date.now())
         .exec((err, user: any) => {
           if (err) { return next(err); }
           if (!user) {
-            req.flash("errors", { msg: "Password reset token is invalid or has expired." });
-            return res.redirect("back");
+            req.flash('errors', { msg: 'Password reset token is invalid or has expired.' });
+            return res.redirect('back');
           }
           user.password = req.body.password;
           user.passwordResetToken = undefined;
@@ -279,7 +279,7 @@ export let postReset = (req: Request, res: Response, next: NextFunction) => {
     },
     function sendResetPasswordEmail(user: UserModel, done: Function) {
       const transporter = nodemailer.createTransport({
-        service: "SendGrid",
+        service: 'SendGrid',
         auth: {
           user: process.env.SENDGRID_USER,
           pass: process.env.SENDGRID_PASSWORD
@@ -287,18 +287,18 @@ export let postReset = (req: Request, res: Response, next: NextFunction) => {
       });
       const mailOptions = {
         to: user.email,
-        from: "express-ts@starter.com",
-        subject: "Your password has been changed",
+        from: 'express-ts@starter.com',
+        subject: 'Your password has been changed',
         text: `Hello,\n\nThis is a confirmation that the password for your account ${user.email} has just been changed.\n`
       };
       transporter.sendMail(mailOptions, (err) => {
-        req.flash("success", { msg: "Success! Your password has been changed." });
+        req.flash('success', { msg: 'Success! Your password has been changed.' });
         done(err);
       });
     }
   ], (err) => {
     if (err) { return next(err); }
-    res.redirect("/");
+    res.redirect('/');
   });
 };
 
@@ -308,10 +308,10 @@ export let postReset = (req: Request, res: Response, next: NextFunction) => {
  */
 export let getForgot = (req: Request, res: Response) => {
   if (req.isAuthenticated()) {
-    return res.redirect("/");
+    return res.redirect('/');
   }
-  res.render("account/forgot", {
-    title: "Forgot Password"
+  res.render('account/forgot', {
+    title: 'Forgot Password'
   });
 };
 
@@ -320,20 +320,20 @@ export let getForgot = (req: Request, res: Response) => {
  * Create a random token, then the send user an email with a reset link.
  */
 export let postForgot = (req: Request, res: Response, next: NextFunction) => {
-  req.assert("email", "Please enter a valid email address.").isEmail();
-  req.sanitize("email").normalizeEmail({ gmail_remove_dots: false });
+  req.assert('email', 'Please enter a valid email address.').isEmail();
+  req.sanitize('email').normalizeEmail({ gmail_remove_dots: false });
 
   const errors = req.validationErrors();
 
   if (errors) {
-    req.flash("errors", errors);
-    return res.redirect("/forgot");
+    req.flash('errors', errors);
+    return res.redirect('/forgot');
   }
 
   async.waterfall([
     function createRandomToken(done: Function) {
       crypto.randomBytes(16, (err, buf) => {
-        const token = buf.toString("hex");
+        const token = buf.toString('hex');
         done(err, token);
       });
     },
@@ -341,8 +341,8 @@ export let postForgot = (req: Request, res: Response, next: NextFunction) => {
       User.findOne({ email: req.body.email }, (err, user: any) => {
         if (err) { return done(err); }
         if (!user) {
-          req.flash("errors", { msg: "Account with that email address does not exist." });
-          return res.redirect("/forgot");
+          req.flash('errors', { msg: 'Account with that email address does not exist.' });
+          return res.redirect('/forgot');
         }
         user.passwordResetToken = token;
         user.passwordResetExpires = Date.now() + 3600000; // 1 hour
@@ -353,7 +353,7 @@ export let postForgot = (req: Request, res: Response, next: NextFunction) => {
     },
     function sendForgotPasswordEmail(token: AuthToken, user: UserModel, done: Function) {
       const transporter = nodemailer.createTransport({
-        service: "SendGrid",
+        service: 'SendGrid',
         auth: {
           user: process.env.SENDGRID_USER,
           pass: process.env.SENDGRID_PASSWORD
@@ -361,20 +361,20 @@ export let postForgot = (req: Request, res: Response, next: NextFunction) => {
       });
       const mailOptions = {
         to: user.email,
-        from: "hackathon@starter.com",
-        subject: "Reset your password on Hackathon Starter",
+        from: 'hackathon@starter.com',
+        subject: 'Reset your password on Hackathon Starter',
         text: `You are receiving this email because you (or someone else) have requested the reset of the password for your account.\n\n
           Please click on the following link, or paste this into your browser to complete the process:\n\n
           http://${req.headers.host}/reset/${token}\n\n
           If you did not request this, please ignore this email and your password will remain unchanged.\n`
       };
       transporter.sendMail(mailOptions, (err) => {
-        req.flash("info", { msg: `An e-mail has been sent to ${user.email} with further instructions.` });
+        req.flash('info', { msg: `An e-mail has been sent to ${user.email} with further instructions.` });
         done(err);
       });
     }
   ], (err) => {
     if (err) { return next(err); }
-    res.redirect("/forgot");
+    res.redirect('/forgot');
   });
 };
