@@ -1,11 +1,11 @@
-import { Request, Response } from "express";
+import { Request, Response } from 'express';
 
 /**
  * GET /
  * Home page.
  */
 export let index = (req: Request, res: Response) => {
-  res.render("home", {
-    title: "Home"
+  res.render('home', {
+    title: 'Home'
   });
 };
