@@ -1,6 +1,6 @@
-import errorHandler from "errorhandler";
+import errorHandler from 'errorhandler';
 
-import app from "./app";
+import app from './app';
 
 /**
  * Error Handler. Provides full stack - remove for production
@@ -10,13 +10,13 @@ app.use(errorHandler());
 /**
  * Start Express server.
  */
-const server = app.listen(app.get("port"), () => {
+const server = app.listen(app.get('port'), () => {
   console.log(
-    "  App is running at http://localhost:%d in %s mode",
-    app.get("port"),
-    app.get("env")
+    '  App is running at http://localhost:%d in %s mode',
+    app.get('port'),
+    app.get('env')
   );
-  console.log("  Press CTRL-C to stop\n");
+  console.log('  Press CTRL-C to stop\n');
 });
 
 export default server;
