@@ -1,5 +1,5 @@
 
-/// <reference types="express" />
+/// <reference types='express' />
 
 // Add RequestValidation Interface on to Express's Request Interface.
 declare namespace Express {
@@ -10,5 +10,5 @@ interface Flash {
     flash(type: string, message: any): void;
 }
 
-declare module "express-flash";
+declare module 'express-flash';
 
