@@ -1,16 +1,16 @@
-import winston from "winston";
-import { Logger } from "winston";
-import { ENVIRONMENT } from "./secrets";
+import winston from 'winston';
+import { Logger } from 'winston';
+import { ENVIRONMENT } from './secrets';
 
 const logger = new (Logger)({
     transports: [
-        new (winston.transports.Console)({ level: process.env.NODE_ENV === "production" ? "error" : "debug" }),
-        new (winston.transports.File)({ filename: "debug.log", level: "debug"})
+        new (winston.transports.Console)({ level: process.env.NODE_ENV === 'production' ? 'error' : 'debug' }),
+        new (winston.transports.File)({ filename: 'debug.log', level: 'debug'})
     ]
 });
 
-if (process.env.NODE_ENV !== "production") {
-    logger.debug("Logging initialized at debug level");
+if (process.env.NODE_ENV !== 'production') {
+    logger.debug('Logging initialized at debug level');
 }
 
 export default logger;
